feat: refactor component imports, format
