+++ components/challenge-dialog.tsx
@@ -10,7 +10,7 @@ import {
   DialogContent,
   DialogPortal,
 } from "@/components/ui/dialog";
-import { HackedText } from "@/components/hacked-text";
+import { HackedText } from "@/components/ui/hacked-text";
 import { Textarea } from "@/components/ui/textarea";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
