+++ components/ui/hacked-text.tsx
